refactor(admin): extract Category management into standalone component
Move the category management feature out of the monolithic AdminComponent into a
standalone OnPush AdminCategoryManagerComponent (selector app-admin-category-manager):
the category CRUD (create/update/delete), the drag + arrow reorder, the per-category
parent / low-stock-threshold / tax-group editors, the RO/EN translation editor, and
the two-step add-category wizard, plus all category-management-only local state.

The category template block is moved verbatim (byte-identical) into the child;
:host { display: contents } keeps its rows participating in the parent settings-card
grid so the panel renders unchanged.

The categories list stays owned/loaded by the parent AdminComponent (it is read by
the product form and the page/home content builders and is loaded for both the pages
and settings sections), so it is threaded through a two-way [(categories)] binding;
the panel emits categoriesChange whenever it mutates the list (add / delete / reorder)
so those cross-feature readers stay in sync. In-place edits of individual category
objects (parent / threshold / tax-group) are seen automatically via the shared
references. taxGroups is threaded in read-only (owned by the Settings > Taxes panel
that stays on the parent) to populate the per-category tax-group select.
TranslateService + ToastService stay injected. Behavioural specs move with the code
from the parent spec suites into admin-category-manager.component.spec.ts.
