+++ frontend/src/app/pages/admin/admin.component.branchfill.spec.ts
@@ -2617,172 +2617,6 @@ describe('AdminComponent — branch fill', () => {
     });
   });
 
-  describe('category management', () => {
-    function cat(over: Record<string, unknown> = {}): any {
-      return {
-        id: 'id',
-        slug: 'slug',
-        name: 'name',
-        sort_order: 0,
-        parent_id: null,
-        low_stock_threshold: null,
-        tax_group_id: null,
-        ...over,
-      };
-    }
-
-    it('categoryDescendantIds resolves nested children, ignoring cycles', () => {
-      c.categories = [
-        cat({ id: 'root', slug: 'root' }),
-        cat({ id: 'a', slug: 'a', parent_id: 'root' }),
-        cat({ id: 'b', slug: 'b', parent_id: 'root' }),
-        cat({ id: 'c', slug: 'c', parent_id: 'a' }),
-        cat({ id: 'self', slug: 'self', parent_id: 'self' }), // cycle entry
-      ];
-      const ids = c.categoryDescendantIds('root');
-      expect(ids.has('a')).toBe(true);
-      expect(ids.has('c')).toBe(true);
-      expect(ids.has('root')).toBe(false);
-    });
-
-    it('categoryParentLabel resolves none/found/missing parents', () => {
-      c.categories = [
-        cat({ id: 'p', slug: 'p', name: 'Parent' }),
-        cat({ id: 'child', slug: 'child', parent_id: 'p' }),
-      ];
-      expect(c.categoryParentLabel(cat({ parent_id: null }))).toBe('adminUi.categories.parentNone');
-      expect(c.categoryParentLabel(cat({ parent_id: 'p' }))).toBe('Parent');
-      expect(c.categoryParentLabel(cat({ parent_id: 'gone' }))).toBe(
-        'adminUi.categories.parentNone',
-      );
-    });
-
-    it('categoryParentOptions excludes self and descendants and sorts by name', () => {
-      c.categories = [
-        cat({ id: 'root', slug: 'root', name: 'B' }),
-        cat({ id: 'child', slug: 'child', name: 'A', parent_id: 'root' }),
-        cat({ id: 'other', slug: 'other', name: undefined }),
-      ];
-      const opts = c.categoryParentOptions(cat({ id: 'root', slug: 'root' }));
-      expect(opts.map((o: any) => o.id)).not.toContain('root');
-      expect(opts.map((o: any) => o.id)).not.toContain('child');
-      expect(opts.map((o: any) => o.id)).toContain('other');
-    });
-
-    it('moveCategory ignores out-of-range moves and reorders/persists otherwise', () => {
-      c.categories = [
-        cat({ slug: 'a', sort_order: 0 }),
-        cat({ slug: 'b' /* missing sort_order */ }),
-        cat({ slug: 'c', sort_order: 2 }),
-      ];
-      c.moveCategory(cat({ slug: 'a' }), -1); // swapIndex < 0
-      c.moveCategory(cat({ slug: 'c', sort_order: 2 }), 1); // swapIndex >= length
-      c.moveCategory(cat({ slug: 'missing' }), 1); // index < 0
-      expect(h.admin.reorderCategories).not.toHaveBeenCalled();
-      h.admin.reorderCategories.and.returnValue(
-        of([cat({ slug: 'b', sort_order: 0 }), cat({ slug: 'a', sort_order: 1 })]),
-      );
-      c.moveCategory(cat({ slug: 'a', sort_order: 0 }), 1);
-      expect(h.toast.success).toHaveBeenCalledWith('adminUi.categories.success.reorder');
-    });
-
-    it('moveCategory reports a reorder error', () => {
-      c.categories = [cat({ slug: 'a', sort_order: 0 }), cat({ slug: 'b', sort_order: 1 })];
-      h.admin.reorderCategories.and.returnValue(throwError(() => ({})));
-      c.moveCategory(cat({ slug: 'a', sort_order: 0 }), 1);
-      expect(h.toast.error).toHaveBeenCalledWith('adminUi.categories.errors.reorder');
-    });
-
-    it('onCategoryDrop ignores empty/identical drags and missing endpoints', () => {
-      c.categories = [cat({ slug: 'a', sort_order: 0 }), cat({ slug: 'b', sort_order: 1 })];
-      c.draggingSlug = null;
-      c.onCategoryDrop('a');
-      c.draggingSlug = 'a';
-      c.onCategoryDrop('a');
-      c.draggingSlug = 'ghost';
-      c.onCategoryDrop('b');
-      expect(h.admin.reorderCategories).not.toHaveBeenCalled();
-    });
-
-    it('onCategoryDrop reorders, persists and clears the dragging slug', () => {
-      c.categories = [
-        cat({ slug: 'a', sort_order: 0 }),
-        cat({ slug: 'b', sort_order: 1 }),
-        cat({ slug: 'c', sort_order: 2 }),
-      ];
-      h.admin.reorderCategories.and.returnValue(of([cat({ slug: 'b', sort_order: 0 })]));
-      c.draggingSlug = 'a';
-      c.onCategoryDrop('c');
-      expect(h.admin.reorderCategories).toHaveBeenCalled();
-      expect(c.draggingSlug).toBeNull();
-    });
-
-    it('onCategoryDrop reports a reorder error', () => {
-      c.categories = [cat({ slug: 'a', sort_order: 0 }), cat({ slug: 'b', sort_order: 1 })];
-      h.admin.reorderCategories.and.returnValue(throwError(() => ({})));
-      c.draggingSlug = 'a';
-      c.onCategoryDrop('b');
-      expect(h.toast.error).toHaveBeenCalledWith('adminUi.categories.errors.reorder');
-    });
-
-    it('updateCategoryLowStockThreshold validates, ignores no-ops, persists and reverts', () => {
-      const bad = cat({ slug: 's', low_stock_threshold: 5 });
-      c.updateCategoryLowStockThreshold(bad, '-3');
-      expect(bad.low_stock_threshold).toBe(5);
-      expect(h.toast.error).toHaveBeenCalledWith(
-        'adminUi.categories.errors.updateLowStockThreshold',
-      );
-
-      const same = cat({ slug: 's', low_stock_threshold: null });
-      c.updateCategoryLowStockThreshold(same, '   ');
-      expect(h.admin.updateCategory).not.toHaveBeenCalled();
-
-      const ok = cat({ slug: 's', low_stock_threshold: null });
-      h.admin.updateCategory.and.returnValue(of({ low_stock_threshold: 7 }));
-      c.updateCategoryLowStockThreshold(ok, '7');
-      expect(ok.low_stock_threshold).toBe(7);
-
-      const fail = cat({ slug: 's', low_stock_threshold: 1 });
-      h.admin.updateCategory.and.returnValue(throwError(() => ({})));
-      c.updateCategoryLowStockThreshold(fail, '9');
-      expect(fail.low_stock_threshold).toBe(1);
-    });
-
-    it('updateCategoryParent ignores no-ops, persists and reverts', () => {
-      const same = cat({ slug: 's', parent_id: null });
-      c.updateCategoryParent(same, '  ');
-      expect(h.admin.updateCategory).not.toHaveBeenCalled();
-
-      const ok = cat({ slug: 's', parent_id: null });
-      h.admin.updateCategory.and.returnValue(of({ parent_id: 'p2' }));
-      c.updateCategoryParent(ok, 'p1');
-      expect(ok.parent_id).toBe('p2');
-
-      const fail = cat({ slug: 's', parent_id: 'old' });
-      h.admin.updateCategory.and.returnValue(throwError(() => ({})));
-      c.updateCategoryParent(fail, 'new');
-      expect(fail.parent_id).toBe('old');
-      expect(h.toast.error).toHaveBeenCalledWith('adminUi.categories.errors.updateParent');
-    });
-
-    it('updateCategoryTaxGroup ignores no-ops, persists and reverts', () => {
-      const same = cat({ slug: 's', tax_group_id: null });
-      c.updateCategoryTaxGroup(same, '');
-      expect(h.admin.updateCategory).not.toHaveBeenCalled();
-
-      const ok = cat({ slug: 's', tax_group_id: null });
-      h.admin.updateCategory.and.returnValue(of({ tax_group_id: 'g2' }));
-      c.updateCategoryTaxGroup(ok, 'g1');
-      expect(ok.tax_group_id).toBe('g2');
-
-      const fail = cat({ slug: 's', tax_group_id: 'old' });
-      h.admin.updateCategory.and.returnValue(throwError(() => ({})));
-      c.updateCategoryTaxGroup(fail, 'new');
-      expect(fail.tax_group_id).toBe('old');
-      expect(h.toast.error).toHaveBeenCalledWith('adminUi.taxes.errors.categoryAssign');
-    });
-  });
-
   describe('onNavigationDrop', () => {
     function link(id: string): any {
       return { id, url: `/u/${id}`, label: { en: id, ro: id } };
@@ -2824,85 +2658,4 @@ describe('AdminComponent — branch fill', () => {
     });
   });
 
-  describe('category branch fill', () => {
-    it('categoryParentOptions sorts nullish-name categories and skips descendants/self', () => {
-      c.categories = [
-        { id: 'a', slug: 'a', name: null, parent_id: null },
-        { id: 'b', slug: 'b', name: 'Beta', parent_id: null },
-        { id: 'c', slug: 'c', name: 'Alpha', parent_id: 'x' },
-      ];
-      const opts = c.categoryParentOptions({ id: 'x' });
-      // 'x' (self) and 'c' (descendant of x) excluded; '' (null name) sorts before 'Beta'.
-      expect(opts.map((o: any) => o.id)).toEqual(['a', 'b']);
-    });
-
-    it('categoryDescendantIds resolves a parent cycle without looping forever', () => {
-      c.categories = [
-        { id: 'p', slug: 'p', name: 'P', parent_id: 'c2' },
-        { id: 'c1', slug: 'c1', name: 'C1', parent_id: 'p' },
-        { id: 'c2', slug: 'c2', name: 'C2', parent_id: 'c1' },
-      ];
-      const opts = c.categoryParentOptions({ id: 'p' });
-      // p plus its descendants (c1, c2 via the cycle) are all excluded.
-      expect(opts).toEqual([]);
-    });
-
-    it('updateCategoryParent clears the parent when raw is nullish and echoes a null response', () => {
-      const cat: any = { slug: 's', parent_id: 'old' };
-      h.admin.updateCategory.and.returnValue(of({ parent_id: null }));
-      c.updateCategoryParent(cat, null as any);
-      expect(h.admin.updateCategory).toHaveBeenCalledWith('s', { parent_id: null });
-      expect(cat.parent_id).toBeNull();
-      expect(h.toast.success).toHaveBeenCalled();
-    });
-
-    it('updateCategoryLowStockThreshold clears the threshold for nullish input and echoes null', () => {
-      const cat: any = { slug: 's', low_stock_threshold: 5 };
-      h.admin.updateCategory.and.returnValue(of({ low_stock_threshold: null }));
-      c.updateCategoryLowStockThreshold(cat, null as any);
-      expect(h.admin.updateCategory).toHaveBeenCalledWith('s', { low_stock_threshold: null });
-      expect(cat.low_stock_threshold).toBeNull();
-    });
-
-    it('updateCategoryTaxGroup clears the group when raw is nullish and echoes null', () => {
-      const cat: any = { slug: 's', tax_group_id: 'g1' };
-      h.admin.updateCategory.and.returnValue(of({ tax_group_id: null }));
-      c.updateCategoryTaxGroup(cat, null as any);
-      expect(h.admin.updateCategory).toHaveBeenCalledWith('s', { tax_group_id: null });
-      expect(cat.tax_group_id).toBeNull();
-    });
-
-    it('moveCategory reorders categories whose sort_order is undefined', () => {
-      c.categories = [
-        { slug: 'a', sort_order: undefined },
-        { slug: 'b', sort_order: undefined },
-      ];
-      h.admin.reorderCategories.and.returnValue(of([{ slug: 'b' }, { slug: 'a' }]));
-      c.moveCategory(c.categories[0], 1);
-      expect(h.admin.reorderCategories).toHaveBeenCalledWith([
-        { slug: 'a', sort_order: 0 },
-        { slug: 'b', sort_order: 0 },
-      ]);
-      expect(c.categories.map((x: any) => x.slug)).toEqual(['b', 'a']);
-      expect(h.toast.success).toHaveBeenCalled();
-    });
-
-    it('onCategoryDrop reorders categories whose sort_order is undefined', () => {
-      c.categories = [
-        { slug: 'a', sort_order: undefined },
-        { slug: 'b', sort_order: undefined },
-      ];
-      c.draggingSlug = 'a';
-      h.admin.reorderCategories.and.returnValue(of([{ slug: 'b' }, { slug: 'a' }]));
-      c.onCategoryDrop('b');
-      expect(h.admin.reorderCategories).toHaveBeenCalled();
-      expect(c.categories.map((x: any) => x.slug)).toEqual(['b', 'a']);
-      expect(c.draggingSlug).toBeNull();
-    });
-
-    it('deleteTaxRate ignores a blank country code', () => {
-      c.deleteTaxRate({ id: 'g' } as any, '   ');
-      expect(h.taxesAdmin.deleteRate).not.toHaveBeenCalled();
-    });
-  });
 });
