+++ frontend/src/app/pages/admin/admin.component.coverage.spec.ts
@@ -1181,61 +1181,6 @@ describe('AdminComponent — orders, users, categories, coupons', () => {
     expect(h.toast.error).toHaveBeenCalled();
   });
 
-  it('moveCategory swaps sort order and respects bounds', () => {
-    c.categories = [
-      { slug: 'a', sort_order: 0 },
-      { slug: 'b', sort_order: 1 },
-    ];
-    h.admin.reorderCategories.and.returnValue(
-      of([
-        { slug: 'b', sort_order: 0 },
-        { slug: 'a', sort_order: 1 },
-      ]),
-    );
-    c.moveCategory({ slug: 'a' } as any, 1);
-    expect(h.admin.reorderCategories).toHaveBeenCalled();
-    expect(h.toast.success).toHaveBeenCalled();
-
-    c.moveCategory({ slug: 'a' } as any, -5); // out of range → no call
-    h.admin.reorderCategories.calls.reset();
-    c.moveCategory({ slug: 'a' } as any, 5);
-    expect(h.admin.reorderCategories).not.toHaveBeenCalled();
-
-    h.admin.reorderCategories.and.returnValue(throwError(() => new Error('x')));
-    c.categories = [
-      { slug: 'a', sort_order: 0 },
-      { slug: 'b', sort_order: 1 },
-    ];
-    c.moveCategory({ slug: 'a' } as any, 1);
-    expect(h.toast.error).toHaveBeenCalled();
-  });
-
-  it('category drag-drop reorders and guards self/missing drops', () => {
-    c.categories = [
-      { slug: 'a', sort_order: 0 },
-      { slug: 'b', sort_order: 1 },
-      { slug: 'c', sort_order: 2 },
-    ];
-    c.onCategoryDragStart('a');
-    expect(c.draggingSlug).toBe('a');
-    const ev = dragEvent();
-    c.onCategoryDragOver(ev);
-    expect(ev.preventDefault).toHaveBeenCalled();
-
-    h.admin.reorderCategories.and.returnValue(of(c.categories));
-    c.onCategoryDrop('c');
-    expect(h.admin.reorderCategories).toHaveBeenCalled();
-    expect(c.draggingSlug).toBeNull();
-
-    c.draggingSlug = 'a';
-    c.onCategoryDrop('a'); // self → reset, no call
-    expect(c.draggingSlug).toBeNull();
-
-    c.draggingSlug = 'ghost';
-    c.onCategoryDrop('a'); // from missing
-    expect(c.draggingSlug).toBeNull();
-  });
-
   it('createCoupon validates the code and reports success/error', () => {
     c.newCoupon = { code: '' };
     c.createCoupon();
@@ -1586,87 +1531,6 @@ describe('AdminComponent — products and stock', () => {
   });
 });
 
-describe('AdminComponent — category wizard', () => {
-  let h: Harness;
-  let c: any;
-  beforeEach(() => {
-    h = createComponent();
-    c = h.component as any;
-  });
-
-  it('start/exit toggle the wizard state', () => {
-    c.startCategoryWizard();
-    expect(c.categoryWizardOpen()).toBe(true);
-    c.exitCategoryWizard();
-    expect(c.categoryWizardOpen()).toBe(false);
-  });
-
-  it('descriptionKey/nextLabelKey/canNext respond to the current step', () => {
-    c.startCategoryWizard();
-    expect(c.categoryWizardDescriptionKey()).toContain('basics');
-    expect(c.categoryWizardNextLabelKey()).toContain('next');
-    expect(c.categoryWizardCanNext()).toBe(false); // no slug yet
-    c.categoryWizardSlug.set('cat');
-    expect(c.categoryWizardCanNext()).toBe(true);
-    c.categoryWizardStep.set(1);
-    expect(c.categoryWizardNextLabelKey()).toContain('done');
-    expect(c.categoryWizardCanNext()).toBe(true); // last step
-    c.exitCategoryWizard();
-    expect(c.categoryWizardCanNext()).toBe(false); // closed
-  });
-
-  it('prev/next/goToStep navigate with guards', () => {
-    c.startCategoryWizard();
-    c.categoryWizardPrev(); // already 0 → no-op
-    expect(c.categoryWizardStep()).toBe(0);
-
-    c.categoryWizardNext(); // cannot advance without slug
-    expect(h.toast.error).toHaveBeenCalled();
-    expect(c.categoryWizardStep()).toBe(0);
-
-    c.categoryWizardSlug.set('cat');
-    c.categoryWizardNext(); // advances to step 1, opens translations
-    expect(c.categoryWizardStep()).toBe(1);
-
-    c.categoryWizardPrev();
-    expect(c.categoryWizardStep()).toBe(0);
-
-    c.categoryWizardNext(); // last step from 1? at step 0 with slug → step 1
-    c.categoryWizardNext(); // at last step → exit
-    expect(c.categoryWizardOpen()).toBe(false);
-
-    c.startCategoryWizard();
-    c.goToCategoryWizardStep(5); // out of range
-    expect(c.categoryWizardStep()).toBe(0);
-    c.categoryWizardSlug.set(null);
-    c.goToCategoryWizardStep(1); // needs slug
-    expect(c.categoryWizardStep()).toBe(0);
-    c.categoryWizardSlug.set('cat');
-    c.goToCategoryWizardStep(1);
-    expect(c.categoryWizardStep()).toBe(1);
-  });
-
-  it('addCategory validates the name and advances the wizard on success', () => {
-    c.categoryName = '';
-    c.addCategory();
-    expect(h.admin.createCategory).not.toHaveBeenCalled();
-
-    c.startCategoryWizard();
-    c.categoryName = 'Shoes';
-    c.categoryParentId = ' p1 ';
-    h.admin.createCategory.and.returnValue(of({ slug: 'shoes' }));
-    c.addCategory();
-    expect(c.categories.length).toBe(1);
-    expect(c.categoryWizardSlug()).toBe('shoes');
-    expect(c.categoryWizardStep()).toBe(1);
-
-    h.admin.createCategory.and.returnValue(throwError(() => new Error('x')));
-    c.categoryName = 'Hats';
-    c.addCategory();
-    expect(h.toast.error).toHaveBeenCalled();
-  });
-});
-
 describe('AdminComponent — tax groups and rates', () => {
   let h: Harness;
   let c: any;
@@ -1769,189 +1633,6 @@ describe('AdminComponent — tax groups and rates', () => {
     expect(h.toast.success).toHaveBeenCalled();
     h.taxesAdmin.deleteRate.and.returnValue(throwError(() => ({})));
     c.deleteTaxRate({ id: 'g1' } as any, 'RO');
-    expect(h.toast.error).toHaveBeenCalled();
-  });
-});
-
-describe('AdminComponent — category hierarchy and translations', () => {
-  let h: Harness;
-  let c: any;
-  beforeEach(() => {
-    h = createComponent();
-    c = h.component as any;
-  });
-
-  it('categoryParentLabel resolves names or the none label', () => {
-    c.categories = [
-      { id: 'p', name: 'Parent' },
-      { id: 'c', name: 'Child', parent_id: 'p' },
-    ];
-    expect(c.categoryParentLabel({ parent_id: 'p' })).toBe('Parent');
-    expect(c.categoryParentLabel({ parent_id: '' })).toContain('parentNone');
-    expect(c.categoryParentLabel({ parent_id: 'missing' })).toContain('parentNone');
-  });
-
-  it('categoryParentOptions excludes self and descendants', () => {
-    c.categories = [
-      { id: 'a', name: 'A' },
-      { id: 'b', name: 'B', parent_id: 'a' },
-      { id: 'cc', name: 'C', parent_id: 'b' },
-      { id: 'd', name: 'D' },
-    ];
-    const ids = c.categoryParentOptions({ id: 'a' }).map((x: any) => x.id);
-    expect(ids).toEqual(['d']); // b and cc are descendants, a is self
-  });
-
-  it('categoryDescendantIds walks the tree', () => {
-    c.categories = [{ id: 'b', parent_id: 'a' }, { id: 'cc', parent_id: 'b' }, { id: 'orphan' }];
-    const set = c.categoryDescendantIds('a');
-    expect(Array.from(set).sort()).toEqual(['b', 'cc']);
-  });
-
-  it('updateCategoryParent optimistically updates and rolls back on error', () => {
-    const cat = { slug: 's', parent_id: null } as any;
-    cat.parent_id = null;
-    c.updateCategoryParent(cat, ' '); // unchanged → no call
-    expect(h.admin.updateCategory).not.toHaveBeenCalled();
-
-    h.admin.updateCategory.and.returnValue(of({ parent_id: 'p2' }));
-    c.updateCategoryParent(cat, 'p2');
-    expect(cat.parent_id).toBe('p2');
-
-    const cat2 = { slug: 's2', parent_id: 'old' } as any;
-    h.admin.updateCategory.and.returnValue(throwError(() => new Error('x')));
-    c.updateCategoryParent(cat2, 'new');
-    expect(cat2.parent_id).toBe('old'); // rolled back
-  });
-
-  it('updateCategoryLowStockThreshold validates and rolls back', () => {
-    const cat = { slug: 's', low_stock_threshold: 5 } as any;
-    c.updateCategoryLowStockThreshold(cat, '-2'); // invalid
-    expect(cat.low_stock_threshold).toBe(5);
-    expect(h.toast.error).toHaveBeenCalled();
-
-    c.updateCategoryLowStockThreshold(cat, '5'); // unchanged
-    expect(h.admin.updateCategory).not.toHaveBeenCalled();
-
-    h.admin.updateCategory.and.returnValue(of({ low_stock_threshold: 3 }));
-    c.updateCategoryLowStockThreshold(cat, '3');
-    expect(cat.low_stock_threshold).toBe(3);
-
-    const cat2 = { slug: 's2', low_stock_threshold: 1 } as any;
-    h.admin.updateCategory.and.returnValue(throwError(() => new Error('x')));
-    c.updateCategoryLowStockThreshold(cat2, '9');
-    expect(cat2.low_stock_threshold).toBe(1);
-  });
-
-  it('updateCategoryTaxGroup updates and rolls back', () => {
-    const cat = { slug: 's', tax_group_id: null } as any;
-    c.updateCategoryTaxGroup(cat, ''); // unchanged
-    expect(h.admin.updateCategory).not.toHaveBeenCalled();
-    h.admin.updateCategory.and.returnValue(of({ tax_group_id: 'g1' }));
-    c.updateCategoryTaxGroup(cat, 'g1');
-    expect(cat.tax_group_id).toBe('g1');
-    const cat2 = { slug: 's2', tax_group_id: 'old' } as any;
-    h.admin.updateCategory.and.returnValue(throwError(() => new Error('x')));
-    c.updateCategoryTaxGroup(cat2, 'g2');
-    expect(cat2.tax_group_id).toBe('old');
-  });
-
-  it('category delete confirm flow opens, confirms and closes', () => {
-    c.openCategoryDeleteConfirm({ slug: 's', id: 'c1' });
-    expect(c.categoryDeleteConfirmOpen()).toBe(true);
-    h.admin.deleteCategory.and.returnValue(of({}));
-    c.categories = [{ slug: 's' }];
-    c.confirmDeleteCategory();
-    expect(c.categoryDeleteConfirmOpen()).toBe(false);
-    expect(c.categories.length).toBe(0);
-
-    // guards: no target, busy
-    c.confirmDeleteCategory();
-    c.categoryDeleteConfirmTarget.set({ slug: 's2' });
-    c.categoryDeleteConfirmBusy.set(true);
-    c.confirmDeleteCategory();
-    expect(h.admin.deleteCategory.calls.count()).toBe(1);
-  });
-
-  it('deleteCategory removes the category and reports failure', () => {
-    c.categories = [{ slug: 'x' }];
-    c.categoryTranslationsSlug = 'x';
-    h.admin.deleteCategory.and.returnValue(of({}));
-    const done = jasmine.createSpy('done');
-    c.deleteCategory('x', { done });
-    expect(done).toHaveBeenCalledWith(true);
-    expect(c.categoryTranslationsSlug).toBeNull();
-
-    h.admin.deleteCategory.and.returnValue(throwError(() => new Error('x')));
-    const done2 = jasmine.createSpy('done2');
-    c.deleteCategory('y', { done: done2 });
-    expect(done2).toHaveBeenCalledWith(false);
-  });
-
-  it('toggleCategoryTranslations opens and closes by slug', () => {
-    h.admin.getCategoryTranslations.and.returnValue(
-      of([{ lang: 'en', name: 'EN', description: 'd' }, { lang: 'fr' }]),
-    );
-    c.toggleCategoryTranslations('s1');
-    expect(c.categoryTranslationsSlug).toBe('s1');
-    expect(c.categoryTranslationExists.en).toBe(true);
-    c.toggleCategoryTranslations('s1'); // toggle off
-    expect(c.categoryTranslationsSlug).toBeNull();
-  });
-
-  it('loadCategoryTranslations handles load errors', () => {
-    h.admin.getCategoryTranslations.and.returnValue(throwError(() => new Error('x')));
-    c.toggleCategoryTranslations('s2');
-    expect(c.categoryTranslationsError()).toBeTruthy();
-  });
-
-  it('saveCategoryTranslation validates name and persists', () => {
-    c.categoryTranslationsSlug = null;
-    c.saveCategoryTranslation('en');
-    expect(h.admin.upsertCategoryTranslation).not.toHaveBeenCalled();
-
-    c.categoryTranslationsSlug = 's1';
-    c.categoryTranslations = {
-      en: { name: '', description: '' },
-      ro: { name: '', description: '' },
-    };
-    c.saveCategoryTranslation('en'); // empty name
-    expect(h.toast.error).toHaveBeenCalled();
-
-    c.categoryTranslations.en = { name: 'Hello', description: ' world ' };
-    h.admin.upsertCategoryTranslation.and.returnValue(of({ name: 'Hello', description: 'world' }));
-    c.saveCategoryTranslation('en');
-    expect(c.categoryTranslationExists.en).toBe(true);
-
-    h.admin.upsertCategoryTranslation.and.returnValue(throwError(() => new Error('x')));
-    c.saveCategoryTranslation('en');
-    expect(c.categoryTranslationsError()).toBeTruthy();
-  });
-
-  it('deleteCategoryTranslation clears the entry and handles errors', () => {
-    c.categoryTranslationsSlug = null;
-    c.deleteCategoryTranslation('ro');
-    expect(h.admin.deleteCategoryTranslation).not.toHaveBeenCalled();
-
-    c.categoryTranslationsSlug = 's1';
-    c.categoryTranslationExists = { en: true, ro: true };
-    h.admin.deleteCategoryTranslation.and.returnValue(of({}));
-    c.deleteCategoryTranslation('ro');
-    expect(c.categoryTranslationExists.ro).toBe(false);
-
-    h.admin.deleteCategoryTranslation.and.returnValue(throwError(() => new Error('x')));
-    c.deleteCategoryTranslation('en');
-    expect(c.categoryTranslationsError()).toBeTruthy();
-  });
-
-  it('duplicateProduct reports success and error', () => {
-    spyOn(c, 'loadAll');
-    spyOn(c, 'loadProduct');
-    h.admin.duplicateProduct.and.returnValue(of({ slug: 'copy' }));
-    c.duplicateProduct('orig');
-    expect(c.loadProduct).toHaveBeenCalledWith('copy');
-    h.admin.duplicateProduct.and.returnValue(throwError(() => new Error('x')));
-    c.duplicateProduct('orig');
     expect(h.toast.error).toHaveBeenCalled();
   });
 });
@@ -5414,17 +5095,6 @@ describe('AdminComponent — remaining edge branches', () => {
     c.contentForm = { title: 'T', body_markdown: 'B', status: 'draft' };
     h.admin.updateContentBlock.and.returnValue(throwError(() => ({ status: 409 })));
     c.saveContent();
-    expect(h.toast.error).toHaveBeenCalled();
-  });
-
-  it('onCategoryDrop reports a reorder failure', () => {
-    c.categories = [
-      { slug: 'a', sort_order: 0 },
-      { slug: 'b', sort_order: 1 },
-    ];
-    c.draggingSlug = 'a';
-    h.admin.reorderCategories.and.returnValue(throwError(() => new Error('x')));
-    c.onCategoryDrop('b');
     expect(h.toast.error).toHaveBeenCalled();
   });
