+++ frontend/src/app/pages/admin/admin.component.spec.ts
@@ -1037,270 +1037,6 @@ describe('AdminComponent', () => {
     });
   });
 
-  describe('category management', () => {
-    it('runs the category wizard navigation', () => {
-      const env = build('settings');
-      env.c.startCategoryWizard();
-      expect(env.c.categoryWizardOpen()).toBeTrue();
-      expect(env.c.categoryWizardCanNext()).toBeFalse();
-      expect(env.c.categoryWizardNextLabelKey()).toBe('adminUi.actions.next');
-      env.c.categoryWizardNext();
-      expect(env.toast.error).toHaveBeenCalled();
-      env.c.categoryWizardSlug.set('slug-1');
-      expect(env.c.categoryWizardCanNext()).toBeTrue();
-      const openTr = spyOn(env.c, 'openCategoryWizardTranslations').and.stub();
-      env.c.categoryWizardNext();
-      expect(env.c.categoryWizardStep()).toBe(1);
-      expect(openTr).toHaveBeenCalled();
-      expect(env.c.categoryWizardNextLabelKey()).toBe('adminUi.actions.done');
-      env.c.categoryWizardPrev();
-      expect(env.c.categoryWizardStep()).toBe(0);
-      env.c.categoryWizardPrev();
-      expect(env.c.categoryWizardStep()).toBe(0);
-      env.c.categoryWizardStep.set(1);
-      env.c.categoryWizardNext();
-      expect(env.c.categoryWizardOpen()).toBeFalse();
-    });
-
-    it('goToCategoryWizardStep validates bounds + slug', () => {
-      const env = build('settings');
-      env.c.goToCategoryWizardStep(0);
-      expect(env.c.categoryWizardOpen()).toBeFalse();
-      env.c.startCategoryWizard();
-      env.c.goToCategoryWizardStep(-1);
-      env.c.goToCategoryWizardStep(99);
-      expect(env.c.categoryWizardStep()).toBe(0);
-      env.c.goToCategoryWizardStep(1);
-      expect(env.toast.error).toHaveBeenCalled();
-      env.c.categoryWizardSlug.set('s');
-      const openTr = spyOn(env.c, 'openCategoryWizardTranslations').and.stub();
-      env.c.goToCategoryWizardStep(1);
-      expect(openTr).toHaveBeenCalled();
-    });
-
-    it('categoryWizardDescriptionKey falls back to basics', () => {
-      const env = build('settings');
-      env.c.categoryWizardStep.set(99);
-      expect(env.c.categoryWizardDescriptionKey()).toBe('adminUi.categories.wizard.desc.basics');
-    });
-
-    it('addCategory validates name then prepends the created category', () => {
-      const env = build('settings');
-      env.c.categoryName = '';
-      env.c.addCategory();
-      expect(env.toast.error).toHaveBeenCalled();
-      env.c.categoryName = 'New';
-      env.c.categories = [{ id: 'old' }] as any;
-      env.admin.createCategory.and.returnValue(of({ id: 'new', slug: 'new-cat' }));
-      env.c.addCategory();
-      expect(env.c.categories[0]).toEqual({ id: 'new', slug: 'new-cat' } as any);
-      expect(env.c.categoryName).toBe('');
-    });
-
-    it('addCategory advances the open wizard to translations', () => {
-      const env = build('settings');
-      env.c.startCategoryWizard();
-      env.c.categoryName = 'New';
-      env.admin.createCategory.and.returnValue(of({ id: 'new', slug: 'new-cat' }));
-      const openTr = spyOn(env.c, 'openCategoryWizardTranslations').and.stub();
-      env.c.addCategory();
-      expect(env.c.categoryWizardSlug()).toBe('new-cat');
-      expect(openTr).toHaveBeenCalled();
-    });
-
-    it('addCategory toasts on failure', () => {
-      const env = build('settings');
-      env.c.categoryName = 'New';
-      env.admin.createCategory.and.returnValue(throwError(() => new Error('x')));
-      env.c.addCategory();
-      expect(env.toast.error).toHaveBeenCalled();
-    });
-
-    it('categoryParentLabel resolves parent names', () => {
-      const env = build('settings');
-      env.c.categories = [
-        { id: 'p', name: 'Parent' },
-        { id: 'c', name: 'Child', parent_id: 'p' },
-      ] as any;
-      expect(env.c.categoryParentLabel({ parent_id: '' } as any)).toBe(
-        'adminUi.categories.parentNone',
-      );
-      expect(env.c.categoryParentLabel({ parent_id: 'p' } as any)).toBe('Parent');
-      expect(env.c.categoryParentLabel({ parent_id: 'missing' } as any)).toBe(
-        'adminUi.categories.parentNone',
-      );
-    });
-
-    it('categoryParentOptions excludes self + descendants', () => {
-      const env = build('settings');
-      env.c.categories = [
-        { id: 'root', name: 'Root' },
-        { id: 'child', name: 'Child', parent_id: 'root' },
-        { id: 'grand', name: 'Grand', parent_id: 'child' },
-        { id: 'other', name: 'Other' },
-      ] as any;
-      const options = env.c.categoryParentOptions({ id: 'root' } as any).map((c) => c.id);
-      expect(options).toEqual(['other']);
-    });
-
-    it('updateCategoryParent skips no-ops and persists changes', () => {
-      const env = build('settings');
-      const cat = { slug: 's', parent_id: 'p' } as any;
-      env.c.updateCategoryParent(cat, 'p');
-      expect(env.admin.updateCategory).not.toHaveBeenCalled();
-      env.admin.updateCategory.and.returnValue(of({ parent_id: 'q' }));
-      env.c.updateCategoryParent(cat, 'q');
-      expect(cat.parent_id).toBe('q');
-    });
-
-    it('updateCategoryParent rolls back on failure', () => {
-      const env = build('settings');
-      const cat = { slug: 's', parent_id: 'p' } as any;
-      env.admin.updateCategory.and.returnValue(throwError(() => new Error('x')));
-      env.c.updateCategoryParent(cat, 'q');
-      expect(cat.parent_id).toBe('p');
-      expect(env.toast.error).toHaveBeenCalled();
-    });
-
-    it('updateCategoryLowStockThreshold validates + persists', () => {
-      const env = build('settings');
-      const cat = { slug: 's', low_stock_threshold: 5 } as any;
-      env.c.updateCategoryLowStockThreshold(cat, '-1');
-      expect(cat.low_stock_threshold).toBe(5);
-      expect(env.toast.error).toHaveBeenCalled();
-      env.c.updateCategoryLowStockThreshold(cat, '5');
-      expect(env.admin.updateCategory).not.toHaveBeenCalled();
-      env.admin.updateCategory.and.returnValue(of({ low_stock_threshold: 8 }));
-      env.c.updateCategoryLowStockThreshold(cat, '8');
-      expect(cat.low_stock_threshold).toBe(8);
-    });
-
-    it('updateCategoryLowStockThreshold rolls back on failure', () => {
-      const env = build('settings');
-      const cat = { slug: 's', low_stock_threshold: 5 } as any;
-      env.admin.updateCategory.and.returnValue(throwError(() => new Error('x')));
-      env.c.updateCategoryLowStockThreshold(cat, '8');
-      expect(cat.low_stock_threshold).toBe(5);
-    });
-
-    it('updateCategoryTaxGroup persists + rolls back', () => {
-      const env = build('settings');
-      const cat = { slug: 's', tax_group_id: 'g1' } as any;
-      env.c.updateCategoryTaxGroup(cat, 'g1');
-      expect(env.admin.updateCategory).not.toHaveBeenCalled();
-      env.admin.updateCategory.and.returnValue(of({ tax_group_id: 'g2' }));
-      env.c.updateCategoryTaxGroup(cat, 'g2');
-      expect(cat.tax_group_id).toBe('g2');
-      env.admin.updateCategory.and.returnValue(throwError(() => new Error('x')));
-      env.c.updateCategoryTaxGroup(cat, 'g3');
-      expect(cat.tax_group_id).toBe('g2');
-    });
-
-    it('category delete confirm flow', () => {
-      const env = build('settings');
-      const cat = { slug: 's' } as any;
-      env.c.openCategoryDeleteConfirm(cat);
-      expect(env.c.categoryDeleteConfirmOpen()).toBeTrue();
-      env.admin.deleteCategory.and.returnValue(of({}));
-      env.c.categories = [{ slug: 's' }, { slug: 't' }] as any;
-      env.c.confirmDeleteCategory();
-      expect(env.c.categories.map((x) => x.slug)).toEqual(['t']);
-      expect(env.c.categoryDeleteConfirmOpen()).toBeFalse();
-    });
-
-    it('confirmDeleteCategory guards on missing target + busy', () => {
-      const env = build('settings');
-      env.c.confirmDeleteCategory();
-      expect(env.admin.deleteCategory).not.toHaveBeenCalled();
-      env.c.categoryDeleteConfirmTarget.set({ slug: 's' } as any);
-      env.c.categoryDeleteConfirmBusy.set(true);
-      env.c.confirmDeleteCategory();
-      expect(env.admin.deleteCategory).not.toHaveBeenCalled();
-    });
-
-    it('deleteCategory toasts + closes translations on failure/success', () => {
-      const env = build('settings');
-      env.c.categoryTranslationsSlug = 's';
-      env.admin.deleteCategory.and.returnValue(of({}));
-      env.c.categories = [{ slug: 's' }] as any;
-      env.c.deleteCategory('s');
-      expect(env.c.categoryTranslationsSlug).toBeNull();
-      env.admin.deleteCategory.and.returnValue(throwError(() => new Error('x')));
-      const done = jasmine.createSpy('done');
-      env.c.deleteCategory('z', { done });
-      expect(done).toHaveBeenCalledWith(false);
-    });
-
-    it('toggleCategoryTranslations opens + closes', () => {
-      const env = build('settings');
-      env.admin.getCategoryTranslations.and.returnValue(
-        of([{ lang: 'en', name: 'N', description: 'D' }]),
-      );
-      env.c.toggleCategoryTranslations('s');
-      expect(env.c.categoryTranslationsSlug).toBe('s');
-      expect(env.c.categoryTranslationExists.en).toBeTrue();
-      env.c.toggleCategoryTranslations('s');
-      expect(env.c.categoryTranslationsSlug).toBeNull();
-    });
-
-    it('saveCategoryTranslation validates + persists', () => {
-      const env = build('settings');
-      env.c.saveCategoryTranslation('en');
-      expect(env.admin.upsertCategoryTranslation).not.toHaveBeenCalled();
-      env.c.categoryTranslationsSlug = 's';
-      env.c.categoryTranslations.en = { name: '', description: '' };
-      env.c.saveCategoryTranslation('en');
-      expect(env.toast.error).toHaveBeenCalled();
-      env.c.categoryTranslations.en = { name: 'Name', description: 'Desc' };
-      env.admin.upsertCategoryTranslation.and.returnValue(
-        of({ name: 'Name', description: 'Desc' }),
-      );
-      env.c.saveCategoryTranslation('en');
-      expect(env.c.categoryTranslationExists.en).toBeTrue();
-    });
-
-    it('saveCategoryTranslation surfaces an error', () => {
-      const env = build('settings');
-      env.c.categoryTranslationsSlug = 's';
-      env.c.categoryTranslations.ro = { name: 'Nume', description: '' };
-      env.admin.upsertCategoryTranslation.and.returnValue(throwError(() => new Error('x')));
-      env.c.saveCategoryTranslation('ro');
-      expect(env.c.categoryTranslationsError()).toBe('adminUi.categories.translations.errors.save');
-    });
-
-    it('deleteCategoryTranslation persists + errors', () => {
-      const env = build('settings');
-      env.c.deleteCategoryTranslation('en');
-      expect(env.admin.deleteCategoryTranslation).not.toHaveBeenCalled();
-      env.c.categoryTranslationsSlug = 's';
-      env.c.categoryTranslationExists.en = true;
-      env.admin.deleteCategoryTranslation.and.returnValue(of({}));
-      env.c.deleteCategoryTranslation('en');
-      expect(env.c.categoryTranslationExists.en).toBeFalse();
-      env.admin.deleteCategoryTranslation.and.returnValue(throwError(() => new Error('x')));
-      env.c.deleteCategoryTranslation('ro');
-      expect(env.c.categoryTranslationsError()).toBe(
-        'adminUi.categories.translations.errors.delete',
-      );
-    });
-
-    it('loadCategoryTranslations maps + errors', () => {
-      const env = build('settings');
-      env.admin.getCategoryTranslations.and.returnValue(
-        of([
-          { lang: 'en', name: 'EN', description: 'd' },
-          { lang: 'fr', name: 'skip' },
-        ]),
-      );
-      (env.c as any).loadCategoryTranslations('s');
-      expect(env.c.categoryTranslationExists.en).toBeTrue();
-      expect(env.c.categoryTranslationExists.ro).toBeFalse();
-      env.admin.getCategoryTranslations.and.returnValue(throwError(() => new Error('x')));
-      (env.c as any).loadCategoryTranslations('s');
-      expect(env.c.categoryTranslationsError()).toBe('adminUi.categories.translations.errors.load');
-    });
-  });
-
   describe('tax groups', () => {
     it('loadTaxGroups normalizes + handles errors', () => {
       const env = build('settings');
@@ -1411,78 +1147,6 @@ describe('AdminComponent', () => {
       env.taxesAdmin.deleteRate.and.returnValue(throwError(() => ({})));
       env.c.deleteTaxRate({ id: 'g' } as any, 'RO');
       expect(env.toast.error).toHaveBeenCalled();
-    });
-  });
-
-  describe('category reorder + drag', () => {
-    it('moveCategory swaps sort order and persists', () => {
-      const env = build('settings');
-      env.c.categories = [
-        { slug: 'a', sort_order: 0 },
-        { slug: 'b', sort_order: 1 },
-      ] as any;
-      env.admin.reorderCategories.and.returnValue(
-        of([
-          { slug: 'b', sort_order: 0 },
-          { slug: 'a', sort_order: 1 },
-        ]),
-      );
-      env.c.moveCategory({ slug: 'a' } as any, 1);
-      expect(env.admin.reorderCategories).toHaveBeenCalled();
-      expect(env.c.categories[0].slug).toBe('b');
-    });
-
-    it('moveCategory ignores out-of-range swaps', () => {
-      const env = build('settings');
-      env.c.categories = [{ slug: 'a', sort_order: 0 }] as any;
-      env.c.moveCategory({ slug: 'a' } as any, -1);
-      expect(env.admin.reorderCategories).not.toHaveBeenCalled();
-    });
-
-    it('moveCategory toasts on failure', () => {
-      const env = build('settings');
-      env.c.categories = [
-        { slug: 'a', sort_order: 0 },
-        { slug: 'b', sort_order: 1 },
-      ] as any;
-      env.admin.reorderCategories.and.returnValue(throwError(() => new Error('x')));
-      env.c.moveCategory({ slug: 'a' } as any, 1);
-      expect(env.toast.error).toHaveBeenCalled();
-    });
-
-    it('drag start/over/drop reorders categories', () => {
-      const env = build('settings');
-      env.c.categories = [
-        { slug: 'a', sort_order: 0 },
-        { slug: 'b', sort_order: 1 },
-      ] as any;
-      env.c.onCategoryDragStart('a');
-      expect(env.c.draggingSlug).toBe('a');
-      const evt = { preventDefault: jasmine.createSpy('pd') } as any;
-      env.c.onCategoryDragOver(evt);
-      expect(evt.preventDefault).toHaveBeenCalled();
-      env.admin.reorderCategories.and.returnValue(
-        of([
-          { slug: 'b', sort_order: 0 },
-          { slug: 'a', sort_order: 1 },
-        ]),
-      );
-      env.c.onCategoryDrop('b');
-      expect(env.admin.reorderCategories).toHaveBeenCalled();
-      expect(env.c.draggingSlug).toBeNull();
-    });
-
-    it('onCategoryDrop short-circuits same/missing targets', () => {
-      const env = build('settings');
-      env.c.onCategoryDrop('x');
-      env.c.draggingSlug = 'a';
-      env.c.onCategoryDrop('a');
-      expect(env.c.draggingSlug).toBeNull();
-      env.c.categories = [{ slug: 'a' }] as any;
-      env.c.draggingSlug = 'a';
-      env.c.onCategoryDrop('ghost');
-      expect(env.c.draggingSlug).toBeNull();
-      expect(env.admin.reorderCategories).not.toHaveBeenCalled();
     });
   });
